Removed wildcard support for package dependencies
Wildcards had several drawbacks. Wildcards were resolved on that Ansible
host which executed the Python module, so only packages known to the
package manager on that host will be matched. Support for wildcards
conflicted with versioned package dependencies such as
"Depends: libc6 (>= 2.2.1)" and alternative package names such as
"Depends: default-mta | mail-transport-agent".

To replace a package wildcard use aptitude to get a list of all packages
which match the wildcard. For example, replace pattern '^libqt5.*-dev$'
with the output of:
  aptitude search '^libqt5.*-dev$' -F "%p"