Fixed date formatting in changelogs of RPM spec files
Headers of changelog entries in RPM spec files begin with
a date in english locale, for example:
  * Fri Oct 01 2021 John Wayne <release@home.arpa> - 1.3.3-7

Previously, Python's strftime function was used to generate
the current date string, but this string used the locale's
names for the weekday and month. Then rpmbuild failed with
  error: bad date in %changelog

Now, Python package babel is used to generate the date
string with english locale. The system locale is ignored.

Ref.:
https://stackoverflow.com/a/32785195/6490710
http://babel.pocoo.org/en/latest/dates.html