Replaced role jm1.common with role defaults
Allow to make decisions, e.g. in roles, not based on what the system's
state currently is but on what it should be according to e.g. inventories:

Role jm1.common defines fact distribution_id based on the host's current
operating system which reflects the current system state. It is defined
with Ansible's set_fact module which has a high priority, e.g. it
overrides variables defined in host_vars and group_vars.

Instead variable distribution_id has been added to role defaults. It is
initialized with the same values as fact distribution_id from role
jm1.common, i.e. it is based on the current operating system of a host.
But as role default variables have a low priority they can be overridden
in group_vars and host_vars.

Role jm1.common has been dropped because it is no longer necessary.